requirements.txt
aiofiles==24.1.0
aiogram==3.17.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
alembic==1.14.1
annotated-types==0.7.0
attrs==25.1.0
autopep8==2.3.2
certifi==2025.1.31
charset-normalizer==3.4.1
contourpy==1.3.1
cycler==0.12.1
fonttools==4.55.3
frozenlist==1.5.0
greenlet==3.1.1
idna==3.10
kiwisolver==1.4.7
magic-filter==1.0.12
Mako==1.3.8
MarkupSafe==3.0.2
matplotlib==3.10.0
multidict==6.1.0
numpy==2.2.0
packaging==24.2
pillow==11.0.0
psycopg2-binary==2.9.10
pycodestyle==2.12.1
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
six==1.17.0
SQLAlchemy==2.0.37
typing_extensions==4.12.2
urllib3==2.3.0
yarl==1.18.3
